:art: refactor : 클래스 패키지 구조 개선
