+++ src/main/java/chess/piece/King.java
@@ -1,6 +1,6 @@
-package piece;
+package chess.piece;
 
-import piece.coordinate.Coordinate;
+import chess.piece.coordinate.Coordinate;
 
 public class King extends Piece {
     public King(Team team, int row, char column) {
