+++ src/main/java/chess/piece/Queen.java
@@ -1,6 +1,6 @@
-package piece;
+package chess.piece;
 
-import piece.coordinate.Coordinate;
+import chess.piece.coordinate.Coordinate;
 
 public class Queen extends Piece {
     public Queen(Team team, int row, char column) {
