+++ src/main/java/chess/piece/Team.java
@@ -1,4 +1,4 @@
-package piece;
+package chess.piece;
 
 public enum Team {
     BLACK,WHITE,EMPTY;
