+++ src/test/java/chess/piece/EmptyTest.java
@@ -1,5 +1,8 @@
-package piece;
+package chess.piece;
 
+import chess.piece.Empty;
+import chess.piece.Piece;
+import chess.piece.Team;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
@@ -8,10 +11,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class BishopTest {
+class EmptyTest {
     @Test
-    void Bishop은_자신의_심볼을_반환한다() {
-        Piece bishop = new Bishop(Team.WHITE, 1, 'a');
-        assertThat(bishop.symbol()).isEqualTo('b');
+    void Empty은_자신의_심볼을_반환한다() {
+        Piece empty = new Empty(Team.WHITE, 1, 'a');
+        assertThat(empty.symbol()).isEqualTo('e');
     }
 }
