+++ src/test/java/chess/piece/KingTest.java
@@ -1,5 +1,8 @@
-package piece;
+package chess.piece;
 
+import chess.piece.King;
+import chess.piece.Piece;
+import chess.piece.Team;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
@@ -8,10 +11,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class PawnTest {
+class KingTest {
     @Test
-    void Pawn은_자신의_심볼을_반환한다() {
-        Piece pawn = new Pawn(Team.WHITE, 1, 'a');
-        assertThat(pawn.symbol()).isEqualTo('p');
+    void King은_자신의_심볼을_반환한다() {
+        Piece king = new King(Team.WHITE, 1, 'a');
+        assertThat(king.symbol()).isEqualTo('k');
     }
 }
