+++ src/test/java/chess/piece/PawnTest.java
@@ -1,5 +1,8 @@
-package piece;
+package chess.piece;
 
+import chess.piece.Pawn;
+import chess.piece.Piece;
+import chess.piece.Team;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
@@ -8,10 +11,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class EmptyTest {
+class PawnTest {
     @Test
-    void Empty은_자신의_심볼을_반환한다() {
-        Piece empty = new Empty(Team.WHITE, 1, 'a');
-        assertThat(empty.symbol()).isEqualTo('e');
+    void Pawn은_자신의_심볼을_반환한다() {
+        Piece pawn = new Pawn(Team.WHITE, 1, 'a');
+        assertThat(pawn.symbol()).isEqualTo('p');
     }
 }
