+++ src/test/java/chess/piece/QueenTest.java
@@ -1,5 +1,8 @@
-package piece;
+package chess.piece;
 
+import chess.piece.Piece;
+import chess.piece.Queen;
+import chess.piece.Team;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
@@ -8,10 +11,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class KnightTest {
+class QueenTest {
     @Test
-    void Knight은_자신의_심볼을_반환한다() {
-        Piece knight = new Knight(Team.WHITE, 1, 'a');
-        assertThat(knight.symbol()).isEqualTo('n');
+    void Queen은_자신의_심볼을_반환한다() {
+        Piece queen = new Queen(Team.WHITE, 1, 'a');
+        assertThat(queen.symbol()).isEqualTo('q');
     }
 }
+
+
+
