+++ src/test/java/chess/piece/coordinate/CoordinateTest.java
@@ -1,5 +1,6 @@
-package piece;
+package chess.piece.coordinate;
 
+import chess.piece.coordinate.Coordinate;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
@@ -8,13 +9,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class QueenTest {
+class CoordinateTest {
     @Test
-    void Queen은_자신의_심볼을_반환한다() {
-        Piece queen = new Queen(Team.WHITE, 1, 'a');
-        assertThat(queen.symbol()).isEqualTo('q');
+    void 좌표는_row와_column값을_받아서_생성한다() {
+        Coordinate coordinate = new Coordinate(1, 'a');
+        assertThat(coordinate).isNotNull();
     }
 }
-
-
-
