+++ src/test/java/chess/piece/coordinate/RowTest.java
@@ -1,5 +1,6 @@
-package piece.coordinate;
+package chess.piece.coordinate;
 
+import chess.piece.coordinate.Row;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.params.ParameterizedTest;
